Added AddressId Specific Add/Edit/Remove Operations

+++ Django_Files/addresses/views.py
@@ -29,9 +29,17 @@ def add_address(req):
     if req.GET.get("E_Key") != password:
         return HttpResponse(f"Error: Incorrect Editing Key")
 
-    required_params = ["addressId", "domainName", "ipv4Address", "port"]
+    required_params = ["domainName", "ipv4Address", "port"]
 
-    newAddress = Addresses()
+    if req.GET.get("addressId") == None:
+        return HttpResponse("Error: Address ID Not Specified")
+
+    if Addresses.objects.filter(addressId=req.GET.get("addressId")).exists():
+        newAddress = Addresses.objects.get(addressId=req.GET.get("addressId"))
+    else:
+        newAddress = Addresses()
+        newAddress.addressId = req.GET.get("addressId")
+
     
     for param in required_params:
         if req.GET.get(param) == None:
@@ -47,19 +55,13 @@ def clear_all(req):
     if req.GET.get("E_Key") != password:
         return HttpResponse(f"Error: Incorrect Editing Key")
 
-    Addresses.objects.all().delete()
-    return HttpResponse("1")
+    if req.GET.get("addressId") == None:
+        Addresses.objects.all().delete()
+        return HttpResponse("1")
+
+    if Addresses.objects.filter(addressId=req.GET.get("addressId")).exists():
+        Addresses.objects.get(addressId=req.GET.get("addressId")).delete()
+        return HttpResponse("1")
 
 
-
-
-
-
-
-
-
-
-
-
-
-        
+    return HttpResponse("0")
